IOS: Move net/ip/top code to its own source files
