+++ Source/Core/Core/CMakeLists.txt
@@ -152,6 +152,7 @@ set(SRCS	ActionReplay.cpp
 			IOS/Network/Net.cpp
 			IOS/Network/Socket.cpp
 			IOS/Network/SSL.cpp
+			IOS/Network/IP/Top.cpp
 			IOS/Network/KD/NetKDRequest.cpp
 			IOS/Network/KD/NetKDTime.cpp
 			IOS/Network/KD/NWC24Config.cpp
