+++ Source/Core/Core/Core.vcxproj
@@ -184,6 +184,7 @@
     <ClCompile Include="IOS\Network\Net.cpp" />
     <ClCompile Include="IOS\Network\Socket.cpp" />
     <ClCompile Include="IOS\Network\SSL.cpp" />
+    <ClCompile Include="IOS\Network\IP\Top.cpp" />
     <ClCompile Include="IOS\Network\KD\NetKDRequest.cpp" />
     <ClCompile Include="IOS\Network\KD\NetKDTime.cpp" />
     <ClCompile Include="IOS\Network\KD\NWC24Config.cpp" />
@@ -425,6 +426,7 @@
     <ClInclude Include="IOS\Network\Net.h" />
     <ClInclude Include="IOS\Network\Socket.h" />
     <ClInclude Include="IOS\Network\SSL.h" />
+    <ClInclude Include="IOS\Network\IP\Top.h" />
     <ClInclude Include="IOS\Network\KD\NetKDRequest.h" />
     <ClInclude Include="IOS\Network\KD\NetKDTime.h" />
     <ClInclude Include="IOS\Network\KD\NWC24Config.h" />
